requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.0.0
aiofiles>=23.0.0
langchain>=0.1.0
langgraph>=0.0.40
langchain-google-genai>=0.0.11